Changed the location of the runtime config file to point to ~/.cache using XDG_CACHE_HOME (if it exists). Also added a dbus service file Created method syncDir() instead of a const
Clubbed all the individual libraries of the syncfw into one libbuteosyncfw library.
This eases the usage of buteo fw that the plugins can bother to link to a single
library.
Fixed the script used for testing

First-cut implementation of simplification of sync profile.
Currently, there are two profile xml files.
1. sync xml
2. service xml, which is a sub-profile of sync xml
The original idea was to separate our any online service from
that of the PC or device sync. But this is a bit over-engineering
and can be optimized.
This patch merges the service and sync profiles and only makes use
of the sync profiles. What this means is that there would not be any
/etc/buteo/profiles/service directory

Signed-off-by: Sateesh Kavuri <sateesh@mobilitas.co>