Implementation of the tool to create a template for a buteo plugin
This is the first cut implementation of a tool that helps create a buteo
sync plugin. The user has to be provide few settings for the plugin that
are provided as part of [client,server,storage]-plugin.cfg. Once the
config file is filled, the tool can be used by running the
'gen_template.py client-plugin.cfg' for a client plugin. This will generate
the complete structure of the plugin after which the developer can
make the necessary modifications to suit their needs. This tool makes use of
'Cheetah' templating library in python and can be installed in Ubuntu using
'sudo apt-get install python-cheetah'

Signed-off-by: Sateesh Kavuri <sateesh.kavuri@gmail.com>